+++ stroom-index-server/src/main/java/stroom/search/server/EventSearchResultHandler.java
@@ -27,91 +27,51 @@ import stroom.util.shared.HasTerminate;
 
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
 
 public class EventSearchResultHandler implements ResultHandler {
     private static final Logger LOGGER = LoggerFactory.getLogger(EventSearchResultHandler.class);
 
-    private final LinkedBlockingQueue<EventRefs> pendingMerges = new LinkedBlockingQueue<>();
-    private final Lock lock = new ReentrantLock();
-
     private volatile EventRefs eventRefs;
 
     @Override
-    public void handle(final Map<CoprocessorKey, Payload> payloadMap, final HasTerminate hasTerminate) {
-        if (payloadMap != null) {
+    public boolean handle(final Map<CoprocessorKey, Payload> payloadMap, final HasTerminate hasTerminate) {
+        boolean partialSuccess = true;
+        if (payloadMap != null && payloadMap.size() > 0) {
+            partialSuccess = false;
             for (final Entry<CoprocessorKey, Payload> entry : payloadMap.entrySet()) {
                 final Payload payload = entry.getValue();
                 if (payload instanceof EventRefsPayload) {
                     final EventRefsPayload eventRefsPayload = (EventRefsPayload) payload;
-                    add(eventRefsPayload.getEventRefs(), hasTerminate);
+                    final boolean success = add(eventRefsPayload.getEventRefs(), hasTerminate);
+                    if (success) {
+                        partialSuccess = true;
+                    }
                 }
             }
         }
+        return partialSuccess;
     }
 
     // Non private for testing purposes.
-    public void add(final EventRefs eventRefs, final HasTerminate hasTerminate) {
+    public boolean add(final EventRefs eventRefs, final HasTerminate hasTerminate) {
+        if (hasTerminate.isTerminated()) {
+            return false;
+        }
+
         if (eventRefs != null) {
-            if (hasTerminate.isTerminated()) {
-                pendingMerges.clear();
-
-            } else {
-                // Add the new queue to the pending merge queue ready for merging.
-                try {
-                    pendingMerges.put(eventRefs);
-                } catch (final InterruptedException | RuntimeException e) {
-                    LOGGER.error(e.getMessage(), e);
-                }
-
-                // Try and merge all of the items on the pending merge queue.
-                tryMergePending(hasTerminate);
-            }
-        }
-    }
-
-    private void tryMergePending(final HasTerminate hasTerminate) {
-        // Only 1 thread will get to do a merge.
-        if (lock.tryLock()) {
+            // Add the new queue to the pending merge queue ready for merging.
             try {
-                mergePending(hasTerminate);
-            } finally {
-                lock.unlock();
-            }
-        } else {
-            LOGGER.trace("Another thread is busy merging, so will let it merge my items");
-        }
-    }
-
-    private void mergePending(final HasTerminate hasTerminate) {
-        EventRefs eventRefs = pendingMerges.poll();
-        while (eventRefs != null) {
-            if (hasTerminate.isTerminated()) {
-                // Clear the queue if we are done.
-                pendingMerges.clear();
-
-            } else {
-                try {
-                    mergeRefs(eventRefs);
-                } catch (final RuntimeException e) {
-                    LOGGER.error(e.getMessage(), e);
-                    throw e;
+                if (this.eventRefs == null) {
+                    this.eventRefs = eventRefs;
+                } else {
+                    this.eventRefs.add(eventRefs);
                 }
+            } catch (final RuntimeException e) {
+                LOGGER.error(e.getMessage(), e);
             }
-
-            // Poll the next item.
-            eventRefs = pendingMerges.poll();
         }
-    }
 
-    private void mergeRefs(final EventRefs newEventRefs) {
-        if (eventRefs == null) {
-            eventRefs = newEventRefs;
-        } else {
-            eventRefs.add(newEventRefs);
-        }
+        return true;
     }
 
     public EventRefs getEventRefs() {
@@ -121,18 +81,5 @@ public class EventSearchResultHandler implements ResultHandler {
     @Override
     public Data getResultStore(final String componentId) {
         return null;
-    }
-
-    @Override
-    public void waitForPendingWork(final HasTerminate hasTerminate) {
-        // This assumes that when this method has been called, all calls to addQueue
-        // have been made, thus we will lock and perform a final merge.
-        lock.lock();
-        try {
-            // Perform final merge.
-            mergePending(hasTerminate);
-        } finally {
-            lock.unlock();
-        }
     }
 }