+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1906** : Improved the way search extractions events are grouped so we can extract more events per stream and therefore improve performance.
+
 * Issue **#1907** : Fixed NPE.